+++ banyand/liaison/http/server.go
@@ -72,35 +72,37 @@ type Server interface {
 }
 
 type server struct {
-	creds        credentials.TransportCredentials
-	l            *logger.Logger
-	clientCloser context.CancelFunc
-	mux          *chi.Mux
-	healthMux    *chi.Mux
-	srv          *http.Server
-	healthSrv    *http.Server
-	stopCh       chan struct{}
-	host         string
-	healthHost   string
-	listenAddr   string
-	healthAddr   string
-	grpcAddr     string
-	keyFile      string
-	certFile     string
-	grpcCert     string
-	port         uint32
-	healthPort   uint32
-	closed       uint32
-	tls          bool
+	creds          credentials.TransportCredentials
+	l              *logger.Logger
+	clientCloser   context.CancelFunc
+	mux            *chi.Mux
+	healthMux      *chi.Mux
+	srv            *http.Server
+	healthSrv      *http.Server
+	stopCh         chan struct{}
+	host           string
+	healthHost     string
+	listenAddr     string
+	healthAddr     string
+	grpcAddr       string
+	grpcHealthAddr string
+	keyFile        string
+	certFile       string
+	grpcCert       string
+	port           uint32
+	healthPort     uint32
+	closed         uint32
+	tls            bool
 }
 
 func (p *server) FlagSet() *run.FlagSet {
 	flagSet := run.NewFlagSet("http")
 	flagSet.StringVar(&p.host, "http-host", "", "listen host for http")
-	flagSet.StringVar(&p.healthHost, "health-http-host", "", "health listen host for http")
+	flagSet.StringVar(&p.healthHost, "http-health-host", "", "health listen host for http")
 	flagSet.Uint32Var(&p.port, "http-port", 17913, "listen port for http")
-	flagSet.Uint32Var(&p.healthPort, "health-http-port", 18913, "health listen port for http")
+	flagSet.Uint32Var(&p.healthPort, "http-health-port", 0, "health listen port for http")
 	flagSet.StringVar(&p.grpcAddr, "http-grpc-addr", "localhost:17912", "http server redirect grpc requests to this address")
+	flagSet.StringVar(&p.grpcHealthAddr, "http-grpc-health-addr", "localhost:17912", "http server redirect grpc health check requests to this address")
 	flagSet.StringVar(&p.certFile, "http-cert-file", "", "the TLS cert file of http server")
 	flagSet.StringVar(&p.keyFile, "http-key-file", "", "the TLS key file of http server")
 	flagSet.StringVar(&p.grpcCert, "http-grpc-cert-file", "", "the grpc TLS cert file if grpc server enables tls")
@@ -181,13 +183,19 @@ func (p *server) Serve() run.StopNotify {
 	} else {
 		opts = append(opts, grpc.WithTransportCredentials(p.creds))
 	}
-	client, err := healthcheck.NewClient(ctx, p.l, p.grpcAddr, opts)
+	client, err := healthcheck.NewClient(ctx, p.l, p.grpcHealthAddr, opts)
 	if err != nil {
 		p.l.Error().Err(err).Msg("Failed to health check client")
 		p.tryClose()
 		return p.stopCh
 	}
-	gwMux := runtime.NewServeMux(runtime.WithMetadata(metadataAnnotator()))
+
+	var gwMux *runtime.ServeMux
+	if p.healthPort != 0 {
+		gwMux = runtime.NewServeMux(runtime.WithMetadata(metadataAnnotator()))
+	} else {
+		gwMux = runtime.NewServeMux(runtime.WithMetadata(metadataAnnotator()), runtime.WithHealthzEndpoint(client))
+	}
 	gwHealthMux := runtime.NewServeMux(runtime.WithMetadata(metadataAnnotator()), runtime.WithHealthzEndpoint(client))
 	err = multierr.Combine(
 		commonv1.RegisterServiceHandlerFromEndpoint(ctx, gwMux, p.grpcAddr, opts),
@@ -224,20 +232,22 @@ func (p *server) Serve() run.StopNotify {
 
 		p.tryClose()
 	}()
-	go func() {
-		p.l.Info().Str("healthAddr", p.healthAddr).Msg("Start liaison http health server")
-		var err error
-		if p.tls {
-			err = p.healthSrv.ListenAndServeTLS(p.certFile, p.keyFile)
-		} else {
-			err = p.healthSrv.ListenAndServe()
-		}
-		if err != http.ErrServerClosed {
-			p.l.Error().Err(err)
-		}
+	if p.healthPort != 0 {
+		go func() {
+			p.l.Info().Str("healthAddr", p.healthAddr).Msg("Start liaison http health server")
+			var err error
+			if p.tls {
+				err = p.healthSrv.ListenAndServeTLS(p.certFile, p.keyFile)
+			} else {
+				err = p.healthSrv.ListenAndServe()
+			}
+			if err != http.ErrServerClosed {
+				p.l.Error().Err(err)
+			}
 
-		p.tryClose()
-	}()
+			p.tryClose()
+		}()
+	}
 	return p.stopCh
 }