+++ CHANGES.md
@@ -28,8 +28,9 @@ Release Notes.
 - Add taking the snapshot of data files.
 - Add backup command line tool to backup the data files.
 - Add restore command line tool to restore the data files.
-- Implement Username/Password Authentication for BanyanDB Liaison.
 - Add concurrent barrier to partition merge to improve the performance of the partition merge.
+- Improve the write performance.
+- Add Implement Username/Password Authentication for BanyanDB Liaison.
 
 ### Bug Fixes
 
@@ -61,6 +62,7 @@ Release Notes.
 - Fix metrics system typo.
 - Bump up OAP in CI to 6d262cce62e156bd197177abb3640ea65bb2d38e.
 - Update cespare/xxhash to v2 version.
+- Bump up Go to 1.24.
 
 ### CVEs